Moved admin-server to modules
