+++ modules/admin-server/src/test/kotlin/com/sphereon/oid/fed/server/admin/ApplicationTests.kt
@@ -1,4 +1,4 @@
-package com.sphereon.oid.fed
+package com.sphereon.oid.fed.server.admin
 
 import org.junit.jupiter.api.Test
 import org.springframework.boot.test.context.SpringBootTest
