+++ modules/admin-server/src/test/kotlin/com/sphereon/oid/fed/server/admin/StatusEndpointTest.kt
@@ -1,4 +1,4 @@
-package com.sphereon.oid.fed
+package com.sphereon.oid.fed.server.admin
 
 import org.junit.jupiter.api.Test
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc
